requirements.txt
diffusers
accelerate
Pillow
flask
flask_socketio
flask_cors
